Add files via upload
Authorization when configured is what the user is allowed to do once logged in. You can configure various groups and privileges on the TACACs/RADIUS server that assign various privileges, even what commands they can run from what privilege level etc. Essentially, the device will query the server when a command is executed to see if they have the privileges to execute that command.